Refactor: Move FindMoveHelper to its own file
Extracted the FindMoveHelper class and all move generation logic
from FreeCellSolver.cs into a new file, FindMoveHelper.cs, for
better code organization and maintainability. No functional
changes were made to the move logic. Also updated test data in
FreeCellSolverTests.cs to use a new game state and move history
for split sequence move scenarios.
